Split demo app into Vuetify and Buefy. Add Buefy component styling. Add cron-validator. Readjust validations.
