+++ src/main.ts
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import App from "./App.vue";
-import vuetify from "./plugins/vuetify";
+import vuetify from "./vuetify/vuetify";
 
 Vue.config.productionTip = false;
