+++ src/vuetify/vuetify.ts
@@ -1,6 +1,5 @@
 import Vue from "vue";
 import Vuetify from "vuetify/lib";
-// https://github.com/vuetifyjs/vuetify/issues/5944
 
 Vue.use(Vuetify);
